requirements.txt
# Chess Analysis Project Dependencies

# Core chess library for PGN parsing and Stockfish integration
python-chess>=1.10.0

# Data manipulation and analysis
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Configuration management
pyyaml>=6.0

# Progress bars for batch processing
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Jupyter support (for exploratory work)
jupyter>=1.0.0
ipykernel>=6.25.0

# Optional: Type checking (recommended for development)
mypy>=1.5.0
